Feature bidding management (#78)
* CREATE UPDATE Inbound, BillOfLading And Container

* REMOVE comment

* FIX conflict

* COMMIT for fix other branch

* CRUD bid and biddingDocument

* COMMIT after test

* FIX conflict when merge master

* EDIT logic

* EDIT Logic

* REMOVE update password

* UPDATE Logic

* UPDATE logic

* EDIT function in Reposotories and ADD Bid logic

* EDIT Convention

* EDIT get Combined by User logic

* UPDATE creat, update and edit bid logic

* UPDATE create, update and edit bid logic

* UPDATE biddoc and bid logic

* CRUD Contract, Evidence, Payment, Update Bid Logic

* CRUD Feedback and Report

* DELETE file redundancy

* Evidence impl abtract method

* Combined and contract update

* UPDATE combined and contract logic

* UPDATE Combined Logic

+++ src/main/java/com/crm/models/dto/ContractDto.java
@@ -13,7 +13,7 @@ public class ContractDto {
 
   private Long Id;
 
-  private Integer finesAgainstContractViolation;
+  private Double finesAgainstContractViolation;
 
   private Boolean required;
